Use span from trace context, not parsed header
Replacing a temporary cheat, logs should be correlated to respective
trace spans as propogated from Open Telemetry.

Providing a custom wrapper around ctxlogrus to read the trace out at the
log entry extraction, and place into the resulting entry as a field, as
the only context provided to logrus is optional for use in Hooks.